+++ cmd/gitserver/server/customfetch_test.go
@@ -2,6 +2,9 @@ package server
 
 import (
 	"context"
+	"encoding/json"
+	"io/ioutil"
+	"os"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -11,10 +14,6 @@ import (
 )
 
 func TestEmptyCustomGitFetch(t *testing.T) {
-	customGitFetch = func() map[string][]string {
-		return buildCustomFetchMappings(nil)
-	}
-
 	remoteURL, _ := vcs.ParseURL("git@github.com:sourcegraph/sourcegraph.git")
 	customCmd := customFetchCmd(context.Background(), remoteURL)
 	if customCmd != nil {
@@ -23,6 +22,9 @@ func TestEmptyCustomGitFetch(t *testing.T) {
 }
 
 func TestCustomGitFetch(t *testing.T) {
+	// mock value for test
+	customGitFetchCmdConf = "/tmp/customGitConfig.json"
+
 	mappings := []*schema.CustomGitFetchMapping{
 		{
 			DomainPath: "github.com/foo/normal/one",
@@ -41,6 +43,21 @@ func TestCustomGitFetch(t *testing.T) {
 			Fetch:      "/foo/bar/git fetch things",
 		},
 	}
+
+	c, err := json.Marshal(mappings)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// write config file to disk
+	err = ioutil.WriteFile(customGitFetchCmdConf, c, 0644)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	t.Cleanup(func() {
+		os.Remove(customGitFetchCmdConf)
+	})
 
 	tests := []struct {
 		url          string
@@ -66,10 +83,6 @@ func TestCustomGitFetch(t *testing.T) {
 		},
 	}
 
-	customGitFetch = func() map[string][]string {
-		return buildCustomFetchMappings(mappings)
-	}
-
 	for _, test := range tests {
 		remoteURL, _ := vcs.ParseURL(test.url)
 		customCmd := customFetchCmd(context.Background(), remoteURL)
@@ -82,4 +95,5 @@ func TestCustomGitFetch(t *testing.T) {
 			t.Errorf("URL %q: %v", test.url, diff)
 		}
 	}
+
 }